+++ MODELO_DATOS/MODELO_REQUISICION/LISTAS_API/CENTRO_TRABAJO.cs
@@ -7,7 +7,7 @@ using System.Threading.Tasks;
 namespace MODELO_DATOS.MODELO_REQUISICION.LISTAS_API
 {
    public class CENTRO_TRABAJO {
-        public decimal COD_CENTRO_TRABAJO { get; set; }
+        public int COD_CENTRO_TRABAJO { get; set; }
         public string NOMBRE { get; set; }
     }
 }
