+++ MODELO_DATOS/MODELO_REQUISICION/LISTAS_API/DIA_LABORAL.cs
@@ -7,7 +7,7 @@ using System.Threading.Tasks;
 namespace MODELO_DATOS.MODELO_REQUISICION.LISTAS_API
 {
   public class DIA_LABORAL{
-        public decimal COD_DIAS_LABORAL { get; set; }
+        public int COD_DIAS_LABORAL { get; set; }
         public string NOMBRE { get; set; }
     }
 }
