+++ MODELO_DATOS/MODELO_REQUISICION/LISTAS_API/EQUIPO_VENTA.cs
@@ -7,7 +7,7 @@ using System.Threading.Tasks;
 namespace MODELO_DATOS.MODELO_REQUISICION.LISTAS_API
 {
    public class EQUIPO_VENTA{
-        public decimal COD_EQUIPO_VENTA { get; set; }
+        public int COD_EQUIPO_VENTA { get; set; }
         public string NOMBRE { get; set; }
     }
 }
