+++ MODELO_DATOS/MODELO_REQUISICION/LISTAS_API/GERENCIA.cs
@@ -8,7 +8,7 @@ namespace MODELO_DATOS.MODELO_REQUISICION.LISTAS_API
 {
    public class GERENCIA
     {
-        public decimal COD_GERENCIA { get; set; }
+        public int COD_GERENCIA { get; set; }
         public string NOMBRE { get; set; }
     }
 }
