+++ MODELO_DATOS/MODELO_REQUISICION/LISTAS_API/JORNADA_TRABAJO.cs
@@ -8,7 +8,7 @@ namespace MODELO_DATOS.MODELO_REQUISICION.LISTAS_API
 {
   public  class JORNADA_TRABAJO
     {
-        public decimal COD_JORNADA_TRABAJO { get; set; }
+        public int COD_JORNADA_TRABAJO { get; set; }
         public string NOMBRE { get; set; }
     }
 }
