+++ MODELO_DATOS/MODELO_REQUISICION/LISTAS_API/LIST_CARGO.cs
@@ -8,7 +8,7 @@ namespace MODELO_DATOS.MODELO_REQUISICION.LISTAS_API
 {
    public class LIST_CARGO
     {
-        public decimal COD_CARGO { get; set; }
+        public int COD_CARGO { get; set; }
         public string NOMBRE { get; set; }
     }
 }
