+++ MODELO_DATOS/MODELO_REQUISICION/LISTAS_API/TIPO_SALARIO.cs
@@ -7,7 +7,7 @@ using System.Threading.Tasks;
 namespace MODELO_DATOS.MODELO_REQUISICION.LISTAS_API
 {
   public class TIPO_SALARIO{
-        public decimal COD_TIPO_SALARIO { get; set; }
+        public int COD_TIPO_SALARIO { get; set; }
         public string NOMBRE { get; set; }
     }
 }
