+++ MODELO_DATOS/MODELO_REQUISICION/LISTAS_API/UBICACION_FISICA.cs
@@ -7,7 +7,7 @@ using System.Threading.Tasks;
 namespace MODELO_DATOS.MODELO_REQUISICION.LISTAS_API
 {
   public  class UBICACION_FISICA {
-        public decimal COD_UBICACION_FISICA { get; set; }
+        public int COD_UBICACION_FISICA { get; set; }
         public string NOMBRE { get; set; }
     }
 }
